feat(step3): input, output담당 struct 생성
기존 Racing과 Car에 있던 input, output 메서드를 분리하기 위해 담당 struct를 만들어서 적용했습니다.
